Changed logic in LDAP role a bit
Added check to see if the eduPerson schema has aleady been added. If so,
do not attempt to add it again, as this leads to ldap reporting an
error. Also removed the when on downloadin, unpacking and copying the
eduPerson schema, as this can always be done. If the copying does not
succeed, it is because the same files are already at their destination
and all is well. Before adding the eduPerson schema, it is checked if it
is already present. We cannot remove or replace schemas, without
starting with a clean slate. Therefore, once added it stays as it is.
This inability to change schemas in openLDAP, negates the when checks
for downloading, unpacking and copying the eduPerson schema.