Add support for COmanage develop branch
Getting the develop branch installed must be done in a slightly
different manner. Mainly, it cannot be fetch from GitHib in the same way
you get a specific version. This needs to be done manually instead.
Playbook has been adjusted for this and is able to cope with the
'develop' version as a special case.