Remove the UUID from the metadata
The uuid was added to the metadata in order to grantee a unique
entityID. Normally, a URL is used as that should be unique. In case the
COmanage portal is installed on a private VM, i.e. on a user's local
machine, the URL might not be unique and hence the original inclusion of
the uuid. This feature is unwanted on the develop or master branch and
hence the deletion.