Add uuid when defined
The sp_random_part can be used to add a UUID to the EntityID of the
metadata. Usually this should not be necessary, as the EntityID should
be unique when the FQDN is resolvable. However, when using local VMs,
one has to choose a fake host and domain name. It is not guaranteed that
this EntityID will be unique, hence the additional random part.